backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
apscheduler==3.10.4
stripe==7.8.0
python-dotenv==1.0.0
numpy==1.26.2
scipy==1.11.4
redis==5.0.1
slowapi==0.1.9
